Use static_cast for array allocation to ensure type safety

+++ base/domain/variable2d.cpp
@@ -76,10 +76,10 @@ void Variable2D::set_center_field(Domain2DUniform* s, field2& f)
 
     field_map[s] = &f;
 
-    buffer_map[s][LocationType::XNegative] = new double[s->ny];
-    buffer_map[s][LocationType::XPositive] = new double[s->ny];
-    buffer_map[s][LocationType::YNegative] = new double[s->nx];
-    buffer_map[s][LocationType::YPositive] = new double[s->nx];
+    buffer_map[s][LocationType::XNegative] = new double[static_cast<std::size_t>(s->ny)]();
+    buffer_map[s][LocationType::XPositive] = new double[static_cast<std::size_t>(s->ny)]();
+    buffer_map[s][LocationType::YNegative] = new double[static_cast<std::size_t>(s->nx)]();
+    buffer_map[s][LocationType::YPositive] = new double[static_cast<std::size_t>(s->nx)]();
 
     position_type = VariablePositionType::Center;
 }
@@ -92,10 +92,10 @@ void Variable2D::set_x_edge_field(Domain2DUniform* s, field2& f)
 
     field_map[s] = &f;
 
-    buffer_map[s][LocationType::XNegative] = new double[s->ny];
-    buffer_map[s][LocationType::XPositive] = new double[s->ny];
-    buffer_map[s][LocationType::YNegative] = new double[s->nx];
-    buffer_map[s][LocationType::YPositive] = new double[s->nx];
+    buffer_map[s][LocationType::XNegative] = new double[static_cast<std::size_t>(s->ny)]();
+    buffer_map[s][LocationType::XPositive] = new double[static_cast<std::size_t>(s->ny)]();
+    buffer_map[s][LocationType::YNegative] = new double[static_cast<std::size_t>(s->nx)]();
+    buffer_map[s][LocationType::YPositive] = new double[static_cast<std::size_t>(s->nx)]();
 
     position_type = VariablePositionType::XFace;
 }
@@ -108,10 +108,10 @@ void Variable2D::set_y_edge_field(Domain2DUniform* s, field2& f)
 
     field_map[s] = &f;
 
-    buffer_map[s][LocationType::XNegative] = new double[s->ny];
-    buffer_map[s][LocationType::XPositive] = new double[s->ny];
-    buffer_map[s][LocationType::YNegative] = new double[s->nx];
-    buffer_map[s][LocationType::YPositive] = new double[s->nx];
+    buffer_map[s][LocationType::XNegative] = new double[static_cast<std::size_t>(s->ny)]();
+    buffer_map[s][LocationType::XPositive] = new double[static_cast<std::size_t>(s->ny)]();
+    buffer_map[s][LocationType::YNegative] = new double[static_cast<std::size_t>(s->nx)]();
+    buffer_map[s][LocationType::YPositive] = new double[static_cast<std::size_t>(s->nx)]();
 
     position_type = VariablePositionType::YFace;
 }
@@ -124,10 +124,10 @@ void Variable2D::set_corner_field(Domain2DUniform* s, field2& f)
 
     field_map[s] = &f;
 
-    buffer_map[s][LocationType::XNegative] = new double[s->ny + 1];
-    buffer_map[s][LocationType::XPositive] = new double[s->ny + 1];
-    buffer_map[s][LocationType::YNegative] = new double[s->nx + 1];
-    buffer_map[s][LocationType::YPositive] = new double[s->nx + 1];
+    buffer_map[s][LocationType::XNegative] = new double[static_cast<std::size_t>(s->ny + 1)]();
+    buffer_map[s][LocationType::XPositive] = new double[static_cast<std::size_t>(s->ny + 1)]();
+    buffer_map[s][LocationType::YNegative] = new double[static_cast<std::size_t>(s->nx + 1)]();
+    buffer_map[s][LocationType::YPositive] = new double[static_cast<std::size_t>(s->nx + 1)]();
 
     position_type = VariablePositionType::Corner;
 }
@@ -141,10 +141,10 @@ void Variable2D::set_inner_field(Domain2DUniform* s, field2& f)
     field_map[s] = &f;
 
     // Inner field owns four-side buffers (lengths for inner-grid indexing).
-    buffer_map[s][LocationType::XNegative] = new double[s->ny];
-    buffer_map[s][LocationType::XPositive] = new double[s->ny];
-    buffer_map[s][LocationType::YNegative] = new double[s->nx];
-    buffer_map[s][LocationType::YPositive] = new double[s->nx];
+    buffer_map[s][LocationType::XNegative] = new double[static_cast<std::size_t>(s->ny)]();
+    buffer_map[s][LocationType::XPositive] = new double[static_cast<std::size_t>(s->ny)]();
+    buffer_map[s][LocationType::YNegative] = new double[static_cast<std::size_t>(s->nx)]();
+    buffer_map[s][LocationType::YPositive] = new double[static_cast<std::size_t>(s->nx)]();
 
     position_type = VariablePositionType::Center;
 }
@@ -205,7 +205,7 @@ void Variable2D::set_boundary_value(Domain2DUniform* s, LocationType loc, double
     has_boundary_value_map[s][loc] = true;
     if (loc == LocationType::XNegative || loc == LocationType::XPositive)
     {
-        boundary_value_map[s][loc] = new double[s->ny];
+        boundary_value_map[s][loc] = new double[static_cast<std::size_t>(s->ny)]();
         for (int j = 0; j < s->ny; j++)
             boundary_value_map[s][loc][j] = in_value;
 
@@ -214,7 +214,7 @@ void Variable2D::set_boundary_value(Domain2DUniform* s, LocationType loc, double
     }
     else if (loc == LocationType::YNegative || loc == LocationType::YPositive)
     {
-        boundary_value_map[s][loc] = new double[s->nx];
+        boundary_value_map[s][loc] = new double[static_cast<std::size_t>(s->nx)]();
         for (int i = 0; i < s->nx; i++)
             boundary_value_map[s][loc][i] = in_value;
 
@@ -279,7 +279,7 @@ void Variable2D::set_boundary_value(Domain2DUniform* s, LocationType loc, std::f
             j_size = s->ny;
         }
 
-        boundary_value_map[s][loc] = new double[j_size];
+        boundary_value_map[s][loc] = new double[static_cast<std::size_t>(j_size)]();
 
         int i_idx = (loc == LocationType::XNegative) ? 0 : s->nx;
 
@@ -302,7 +302,7 @@ void Variable2D::set_boundary_value(Domain2DUniform* s, LocationType loc, std::f
             i_size = s->nx;
         }
 
-        boundary_value_map[s][loc] = new double[i_size];
+        boundary_value_map[s][loc] = new double[static_cast<std::size_t>(i_size)]();
 
         int j_idx = (loc == LocationType::YNegative) ? 0 : s->ny;
